JDP190903-33: resolving conflicts
