+++ src/main/java/com/kodilla/ecommercee/CartController.java
@@ -1,8 +1,8 @@
 package com.kodilla.ecommercee;
 
-import com.kodilla.ecommercee.domain.CartDto;
-import com.kodilla.ecommercee.domain.UserOrderDto;
-import com.kodilla.ecommercee.domain.ProductDto;
+import com.kodilla.ecommercee.dto.CartDto;
+import com.kodilla.ecommercee.dto.UserOrderDto;
+import com.kodilla.ecommercee.dto.ProductDto;
 import org.springframework.web.bind.annotation.*;
 import java.util.ArrayList;
 import java.util.List;
