+++ src/main/java/com/kodilla/ecommercee/UserController.java
@@ -1,6 +1,6 @@
 package com.kodilla.ecommercee;
 
-import com.kodilla.ecommercee.domain.UserDto;
+import com.kodilla.ecommercee.dto.UserDto;
 import org.springframework.web.bind.annotation.*;
 import java.util.ArrayList;
 import java.util.List;
