+++ src/main/java/com/kodilla/ecommercee/dto/ProductDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.dto;
+
+public class ProductDto {
+}
