+++ src/main/java/com/kodilla/ecommercee/dto/UserOrderDto.java
@@ -1,16 +1,19 @@
-package com.kodilla.ecommercee.domain;
+package com.kodilla.ecommercee.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
-public class ProductGroupDto {
-    private String name;
+@NoArgsConstructor
+public class UserOrderDto {
     private long id;
+    private String number;
+    private LocalDate orderDate;
+    private long userId;
     private List<ProductDto> products;
 }
